requirements.txt
cv2
torch
torchvision
Pillow
argparse
timm
scipy
os
pandas